1.2.2
优化API
修复了监听器注解的相关bug (稳定了，快去用)
支持了字符串内嵌函数的嵌套
兼容了高精度数字运算

+++ src/main/kotlin/com/skillw/pouvoir/api/formula/Formula.kt
@@ -4,6 +4,7 @@ import com.skillw.pouvoir.Pouvoir.pouPlaceHolderAPI
 import com.skillw.pouvoir.util.CalculationUtils
 import com.skillw.pouvoir.util.StringUtils.replacement
 import org.bukkit.entity.LivingEntity
+import java.math.BigDecimal
 import java.util.concurrent.ConcurrentHashMap
 
 /**
@@ -22,7 +23,7 @@ class Formula(private val formula: String, private val livingEntity: LivingEntit
         return pouPlaceHolderAPI.replace(livingEntity, formula.replacement(replacements.toMap()))
     }
 
-    fun result(): Double {
+    fun result(): BigDecimal {
         return CalculationUtils.getResult(formula(), livingEntity, replacements)
     }
 }
